+++ walkers/src/extras/images.rs
@@ -2,7 +2,7 @@ use crate::{Plugin, Position};
 use egui::epaint::emath::Rot2;
 use egui::{pos2, Color32, ColorImage, Context, Rect, TextureHandle, TextureId};
 
-/// A image to be drawn on the map.
+/// An image to be drawn on the map.
 pub struct Image {
     /// Geographical position.
     pub position: Position,
@@ -70,8 +70,8 @@ impl Texture {
     /// Construct new texture
     /// ⚠️ Make sure to only call this ONCE for each image, i.e. NOT in your main GUI code.
     /// The call is NOT immediate safe.
-    pub fn new(ctx: Context, uri: &str, img: ColorImage) -> Self {
-        let texture = ctx.load_texture(uri, img.clone(), Default::default());
+    pub fn new(ctx: Context, img: ColorImage) -> Self {
+        let texture = ctx.load_texture("texture", img, Default::default());
 
         Self {
             texture,