+++ demo/src/lib.rs
@@ -208,11 +208,11 @@ mod windows {
             .show(ui.ctx(), |ui| {
                 ui.horizontal(|ui| {
                     if ui.button(RichText::new("➕").heading()).clicked() {
-                        let _ = map_memory.zoom.zoom_in();
+                        let _ = map_memory.zoom_in();
                     }
 
                     if ui.button(RichText::new("➖").heading()).clicked() {
-                        let _ = map_memory.zoom.zoom_out();
+                        let _ = map_memory.zoom_out();
                     }
                 });
             });
@@ -220,7 +220,7 @@ mod windows {
 
     /// When map is "detached", show a windows with an option to go back to my position.
     pub fn go_to_my_position(ui: &Ui, map_memory: &mut MapMemory) {
-        if let Some(position) = map_memory.center_mode.detached() {
+        if let Some(position) = map_memory.detached() {
             Window::new("Center")
                 .collapsible(false)
                 .resizable(false)